pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fluidzero-cli"
version = "0.1.0"
description = "FluidZero platform CLI"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [{name = "Force Platforms Inc"}]
dependencies = [
    "click>=8.1",
    "httpx>=0.27",
    "tabulate>=0.9",
    "rich>=13.0",
    "PyJWT>=2.8",
]

[project.scripts]
fz = "fz_cli.main:cli"

[project.urls]
Homepage = "https://github.com/fluidzero/fz-cli"
Repository = "https://github.com/fluidzero/fz-cli"
Issues = "https://github.com/fluidzero/fz-cli/issues"

[tool.setuptools.packages.find]
where = ["src"]
